Create AnonymousGlobal class.
Move away from AbstractClassResolver the anonymous globals logic to its own class
